🚧refactor : 스페이스 생성 아이콘 로직 분리
