+++ src/components/CreateSpace/SpaceIcon.tsx
@@ -1,13 +1,27 @@
 import { ImageArrType, ImageType } from '@/types';
+import { ReactComponent as PhotoIcon } from '@/assets/svg/photoIcon.svg';
 import { S } from '@/pages/CreateSpace/style';
+import { BasicIcon } from '@/components/Create';
+import { BasicBox } from '@/components/common';
+
 
 const SpaceIcon = ({
   setImageArr,
   inputRef,
+  imageArr,
+  onClickOptionModalOpen,
+  setIsIconModalOpen,
+  isBasicIcon,
 }: {
-  setImageArr: any;
+  setImageArr: React.Dispatch<React.SetStateAction<ImageArrType>>;
   inputRef: any;
+  imageArr: ImageArrType;
+  isBasicIcon: [false] | [true, number];
+  onClickOptionModalOpen: () => void;
+  setIsIconModalOpen: React.Dispatch<React.SetStateAction<[boolean, boolean]>>;
 }) => {
+  const BasicIconArr = BasicIcon();
+
   //파일 변경 함수
   const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     e.preventDefault();
@@ -30,18 +44,38 @@ const SpaceIcon = ({
 
   return (
     <>
-      <S.TitleContainer number={1} required={true}>
-        <div>스페이스 아이콘</div>
-        <input
-          type="file"
-          accept="image/*"
-          style={{ display: 'none' }}
-          onChange={(e) => {
-            handleFileChange(e);
-          }}
-          ref={inputRef}
-        />
-      </S.TitleContainer>
+      <input
+        type="file"
+        accept="image/*"
+        style={{ display: 'none' }}
+        onChange={(e) => {
+          handleFileChange(e);
+        }}
+        ref={inputRef}
+      />
+      {!imageArr.cropImage && !isBasicIcon[0] ? (
+        <S.InputContainer number={1} onClick={onClickOptionModalOpen}>
+          <BasicBox width={160} borderradius={10}>
+            <PhotoIcon />
+          </BasicBox>
+        </S.InputContainer>
+      ) : (
+        <S.InputContainer number={1}>
+          <BasicBox
+            onClick={() => {
+              setIsIconModalOpen([true, false]);
+            }}
+            backgroundImage={
+              isBasicIcon[0]
+                ? BasicIconArr[isBasicIcon[1]]
+                : imageArr.cropImage!
+            }
+            width={160}
+            borderradius={10}
+            color="grey"
+          ></BasicBox>
+        </S.InputContainer>
+      )}
     </>
   );
 };
