--- src/components/CreateSpace/SpaceModal.tsx
@@ -1,3 +1,0 @@
-const SpaceModal = () => {};
-
-export default SpaceModal;
