+++ src/router/Router.tsx
@@ -4,6 +4,7 @@ import { Lazy, PrivateRoute } from '@/router';
 import { Suspense } from 'react';
 import { createBrowserRouter } from 'react-router-dom';
 import Auth from '@/pages/Auth';
+import CreateSpace from '@/pages/CreateSpace';
 import LandingPage from '@/pages/Landing';
 import NotFound from '@/pages/NotFound';
 import { CreatePostSkeleton, CreateSpaceSkeleton } from '@/components/Create';
@@ -11,6 +12,7 @@ import { MainPageSkeleton } from '@/components/Main';
 import { MyPageSkeleton } from '@/components/MyPage';
 import { SelectSpaceSkeleton } from '@/components/SelectSpace';
 import { Loading } from '@/components/common';
+
 
 export const router = createBrowserRouter([
   {
@@ -115,6 +117,10 @@ export const router = createBrowserRouter([
         <LandingPage />
       </Suspense>
     ),
+  },
+  {
+    path: '/createSpace',
+    element: <CreateSpace />,
   },
   {
     path: '/*',
